Fix VehicleSchedules.tsx wrapper and complete schedules feature migration
